% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{cleanHockeyData}
\alias{cleanHockeyData}
\title{Clean the CSV imported hockey data}
\usage{
cleanHockeyData(hockey_data, cleanTeams = TRUE, identifyTies = TRUE,
  listWinnersLosers = TRUE, removeInternational = TRUE, eloResults = TRUE)
}
\arguments{
\item{hockey_data}{the combined CSV data}

\item{cleanTeams}{Whether to clean historical team names & track moving teams (Boolean, default = \code{TRUE})}

\item{identifyTies}{Whether to identify ties & adjust results (Boolean, default = \code{TRUE})}

\item{listWinnersLosers}{Whether to add column for Winners & Losers (Boolean, default = \code{TRUE})}

\item{removeInternational}{Whether to add games between NHL teams and international teams (Boolean, default = \code{TRUE})}

\item{eloResults}{Whether to add a column for results in a [0..1] set (Boolean, default = \code{TRUE})}
}
\value{
Returns a cleaned data frame of hockey scores, with options as selected
}
\description{
Clean the CSV imported hockey data
}
\keyword{internal}

