% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersQH.R
\name{combinePlayerDataFrames.QH}
\alias{combinePlayerDataFrames.QH}
\title{Combine single letter player data frames
Easily combine player data.frames into one single RDS file.}
\usage{
combinePlayerDataFrames.QH(directory = "./data/players/",
  return_data_frame = TRUE, ...)
}
\arguments{
\item{directory}{The directory where data files are stored. Default './data/players/'}

\item{return_data_frame}{Boolean, whether to return data or 'TRUE' when complete}

\item{...}{Additional parameters to pass}
}
\value{
TRUE, or the player data.frame, if successful
}
\description{
Combine single letter player data frames
Easily combine player data.frames into one single RDS file.
}
\keyword{internal}

