% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{combinePlayerDataFrames}
\alias{combinePlayerDataFrames}
\title{Combine single letter player data frames
Easily combine player data.frames into one single RDS file.}
\usage{
combinePlayerDataFrames(directory = "./data/players/",
  return_data_frame = TRUE)
}
\arguments{
\item{directory}{The directory where data files are stored. Default './data/players/'}

\item{return_data_frame}{}
}
\value{
TRUE, or the player data.frame, if successful
}
\description{
Combine single letter player data frames
Easily combine player data.frames into one single RDS file.
}
\keyword{internal}

