% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{flattenTables}
\alias{flattenTables}
\title{Scrape & compile a player's info
Read each table on hockey-reference.com and flatten to single useful tables.
Avoids tables such as last5, awards, etc.
Melts all NHL (regular season and playoff) and non-NHL, recording league and playoff status}
\usage{
flattenTables(tables)
}
\arguments{
\item{tables}{Tables output from \code{XML::readHTMLTables} of the player's page}
}
\value{
single combined stats Table as data.frame
}
\description{
Scrape & compile a player's info
Read each table on hockey-reference.com and flatten to single useful tables.
Avoids tables such as last5, awards, etc.
Melts all NHL (regular season and playoff) and non-NHL, recording league and playoff status
}

