% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{getAndSaveNHLGames}
\alias{getAndSaveNHLGames}
\title{Download and save NHL Games}
\usage{
getAndSaveNHLGames(start = 1918, end = 2017, sleep = 30,
  data_dir = "./data/scores/")
}
\arguments{
\item{start}{Starting season to download. Season is \code{value-1 to value} (eg. start=1918 --> 1917-1918 season)}

\item{end}{Last season to download}

\item{sleep}{Time to sleep between scrapes}

\item{data_dir}{Location to save the csv data}
}
\description{
Download NHL games scores from Hockey-Reference.com
}
\keyword{internal}

