% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{getAndSaveWHAGames}
\alias{getAndSaveWHAGames}
\title{Download and save WHA Games}
\usage{
getAndSaveWHAGames(start = 1973, end = 1979, sleep = 30,
  data_dir = "./data/scores/", ...)
}
\arguments{
\item{start}{Starting season to download. Season is \code{value-1 to value} (eg. start=1973 --> 1972-1973 season)}

\item{end}{Last season to download}

\item{sleep}{Time to sleep between scrapes}

\item{data_dir}{Location to save the csv data}

\item{...}{Additional parameters to pass}
}
\description{
Download WHA games scores from Hockey-Reference.com
}
\keyword{internal}
