% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeRosters.R
\name{getCurrentRosters}
\alias{getCurrentRosters}
\title{Get Current estimated rosters}
\usage{
getCurrentRosters(sleep = 30)
}
\arguments{
\item{sleep}{The amount of time to sleep between requests}
}
\value{
a long data frame with three columns:
\item{Team}{The team}
\item{Player}{The name of the Forward, Defence or Goalie playing}
\item{updateDate}{The date the Team's page was updated}
}
\description{
Get Current estimated rosters
}

