% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getCurrentSeason}
\alias{getCurrentSeason}
\title{Get the current Season 'year code'. From 01 August 2016 to 31 July 2017 will return 2017}
\usage{
getCurrentSeason()
}
\value{
Numeric, 4 digit year
}
\description{
Get the current Season 'year code'. From 01 August 2016 to 31 July 2017 will return 2017
}
