% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersQH.R
\name{getPlayerList.QH}
\alias{getPlayerList.QH}
\title{Get a list of Players
Download the full list of active and past players from Quanthockey.com}
\usage{
getPlayerList.QH(prebuilt = FALSE)
}
\arguments{
\item{prebuilt}{Whether t0 use a prebuilt player list or try a new one.}
}
\value{
a data.frame containing:
\item{Complete}{The complete line that data was scraped from}
\item{BlnNHL}{Whether the player played any time in the NHL}
\item{URL}{Player page URL}
\item{Name}{Player Name}
\item{Active}{Whether the player is currently active}
}
\description{
Get a list of Players
Download the full list of active and past players from Quanthockey.com
}
