% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{getPlayerList}
\alias{getPlayerList}
\title{Get a list of Players
Download the full list of active and past players from Hockey-Reference.com}
\usage{
getPlayerList(sleep = 30, letters = letters)
}
\arguments{
\item{sleep}{The sleep time between scrape requests}

\item{letters}{The letters to scrape player names from. Default: \code{letters}}
}
\value{
a data.frame containing:
\item{Complete}{The complete line that data was scraped from}
\item{BlnNHL}{Whether the player played any time in the NHL}
\item{URL}{Player page URL}
\item{Name}{Player Name}
\item{Active}{Whether the player is currently active}
}
\description{
Get a list of Players
Download the full list of active and past players from Hockey-Reference.com
}

