% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersHR.R
\name{getPlayerStats.HR}
\alias{getPlayerStats.HR}
\title{Get stats for a list of players
A function to get all player stats from a list of players (in the form provided by \code{\link{getPlayerList.HR}})}
\usage{
getPlayerStats.HR(player_list, sleep = 30)
}
\arguments{
\item{player_list}{A player list (data.frame) from \code{\link{getPlayerList.HR}}}

\item{sleep}{Time to sleep between player scrapings}
}
\value{
a list of three data.frames, containing
\item{PlayerStats}{Combined player statistics}
\item{GoalieStats}{Combined goalie statistics}
\item{PlayerMeta}{Meta statistics for all (goalies and players)}
}
\description{
Get stats for a list of players
A function to get all player stats from a list of players (in the form provided by \code{\link{getPlayerList.HR}})
}
