% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{getSchedule}
\alias{getSchedule}
\title{Get season schedule}
\usage{
getSchedule(from_date = Sys.Date(), data_dir = "./data/scores",
  include_playoffs = TRUE, ...)
}
\arguments{
\item{from_date}{The date from which to return an empty schedule}

\item{data_dir}{The location to save the schedule}

\item{include_playoffs}{Whether or not to include this years' plaoyffs, if available}

\item{...}{Additional parameters to pass}
}
\value{
A data.frame of schedule, with scores removed
}
\description{
Get season schedule
}

