% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getURLInternal}
\alias{getURLInternal}
\title{Extra wrapped getURL
\code{\link[RCurl]{getURL}} wrapped in useragent sampling and try/retry, with referer.}
\usage{
getURLInternal(url, referer = "http://www.google.ca")
}
\arguments{
\item{url}{The URL to try download}

\item{referer}{A referer source}
}
\value{
raw url contents output from getURL as the HTTP reply from the server, or try-error if an error has occurred and cannot be recovered from.
}
\description{
Extra wrapped getURL
\code{\link[RCurl]{getURL}} wrapped in useragent sampling and try/retry, with referer.
}
\keyword{internal}
