% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{processPlayerData}
\alias{processPlayerData}
\title{Clean Player Data
This function will process player data, returning clean data frames as a list}
\usage{
processPlayerData(player_data, drop_awards = TRUE)
}
\arguments{
\item{player_data}{The player_data to clean up}

\item{drop_awards}{Whether to drop awards column.}
}
\value{
a list of three cleaned data.frames, containing
\item{PlayerStats}{Combined player statistics}
\item{GoalieStats}{Combined goalie statistics}
\item{PlayerMeta}{Meta statistics for all (goalies and players)}
}
\description{
Clean Player Data
This function will process player data, returning clean data frames as a list
}

