% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{readHockeyData}
\alias{readHockeyData}
\title{Read in Hockey Data csv files.}
\usage{
readHockeyData(data_dir = "./data/scores/",
  nhl_year_list = c(1918:getCurrentSeason()), wha_year_list = c(1973:1979),
  playoffs = TRUE, lastPlayoffs = FALSE, ...)
}
\arguments{
\item{data_dir}{The location of the files}

\item{nhl_year_list}{The NHL years to read in}

\item{wha_year_list}{The WHA years to read in}

\item{playoffs}{Whether to read in playoffs (Boolean, default = \code{TRUE})}

\item{lastPlayoffs}{Whether to read the last listed playoffs. Default = FALSE}

\item{...}{Further parameters to pass to \code{\link{cleanHockeyData}}}
}
\value{
returns a un-cleaned data frame from NHL and WHA score csv files
}
\description{
Read in Hockey Data csv files.
}
\keyword{internal}
