% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{scrapeAdvancedStats}
\alias{scrapeAdvancedStats}
\title{Scrape Advanced Stats}
\usage{
scrapeAdvancedStats(scores = NULL, data_dir = "./data/", sleep = 10)
}
\arguments{
\item{scores}{historical scores to pair to advanced stats}

\item{data_dir}{where to store checkpoint data}

\item{sleep}{number of seconds between scrapes}
}
\value{
a larger data frame.
}
\description{
Scrape Advanced Stats
}
