% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersHR.R
\name{scrapeByAlphabet.HR}
\alias{scrapeByAlphabet.HR}
\title{Scrape players by Alphabet
A function to scrape and save player tables by last name, breaking up the scraping
into each chunk to prevent progress loss by scraping error (HTML error)}
\usage{
scrapeByAlphabet.HR(player_list, letters_to_scrape = letters,
  long_sleep = 120, combine = TRUE, directory = "./data/players/", ...)
}
\arguments{
\item{player_list}{a player list (data.frame) of the type created by \code{\link{getPlayerList.HR}}}

\item{letters_to_scrape}{the letters of last names to scrape (default all letters)}

\item{long_sleep}{The length of time to sleep between letters}

\item{combine}{Whether to combine all player data.frames (a-z) after downloading}

\item{directory}{Where to store data files}

\item{...}{Additional params for getPlayerStats}
}
\value{
True, if successful
}
\description{
Scrape players by Alphabet
A function to scrape and save player tables by last name, breaking up the scraping
into each chunk to prevent progress loss by scraping error (HTML error)
}
\keyword{internal}
