% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersQH.R
\name{scrapeByNumber}
\alias{scrapeByNumber}
\title{Scrape players by Alphabet
A function to scrape and save player tables by last name, breaking up the scraping
into each chunk to prevent progress loss by scraping error (HTML error)}
\usage{
scrapeByNumber(player_list, group_by = 1000, long_sleep = 120,
  combine = TRUE, directory = "./data/players/", start_at = 1,
  save_player_list = TRUE, ...)
}
\arguments{
\item{player_list}{a player list (data.frame) of the type created by \code{\link{getPlayerList.QH}}}

\item{group_by}{The group size to scrape. Default 1000}

\item{long_sleep}{The length of time to sleep between groups}

\item{combine}{Whether to combine all player data.frames (groups) after downloading}

\item{directory}{Where to store data files}

\item{start_at}{The playernumber to start at. Default = 1}

\item{save_player_list}{Whether to save the player list}

\item{...}{Additional params for getPlayerStats}
}
\value{
True, if successful
}
\description{
Scrape players by Alphabet
A function to scrape and save player tables by last name, breaking up the scraping
into each chunk to prevent progress loss by scraping error (HTML error)
}
\keyword{internal}

