% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersHR.R
\name{scrapePlayers.HR}
\alias{scrapePlayers.HR}
\title{Scrape and clean all player data}
\usage{
scrapePlayers.HR(data_dir = "./data/players/", ...)
}
\arguments{
\item{data_dir}{Directory to store data}

\item{...}{Additional parameters to pass}
}
\value{
a list of three cleaned data.frames, containing
\item{PlayerStats}{Combined player statistics}
\item{GoalieStats}{Combined goalie statistics}
\item{PlayerMeta}{Meta statistics for all (goalies and players)}
}
\description{
This is a one-command function to scrape and clean all player data available from Hockey-Reference.com. This takes many hours.
Also saves results to a dated .RDS file.
}
\examples{
\dontrun{scrapePlayers.HR()}
\dontrun{scrapePlayers.HR(sleep=15, long_sleep=180)}
}
