% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersQH.R
\name{scrapePlInfo.QH}
\alias{scrapePlInfo.QH}
\title{Get player info from a page
A function to manage scraping all of the info from Quanthockey.com player pages.
Also perform collection player meta info (name, birth, etc)}
\usage{
scrapePlInfo.QH(url)
}
\arguments{
\item{url}{The player URL}
}
\value{
a list of data.frames containing
\item{Tables}{Player Statistics Tables from HTML Page}
\item{Metas}{Player Meta information (Name, Birth, Height, Weight, Shot, etc.)}
}
\description{
Get player info from a page
A function to manage scraping all of the info from Quanthockey.com player pages.
Also perform collection player meta info (name, birth, etc)
}
\keyword{internal}

