% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{scrapePlInfo}
\alias{scrapePlInfo}
\title{Get player info from a page
A function to manage scraping all of the info from Hockey-Reference.com player pages.
Also perform collection player meta info (name, birth, etc)}
\usage{
scrapePlInfo(url)
}
\arguments{
\item{url}{The player URL}
}
\value{
a list of data.frames containing
\item{Tables}{Player Statistics Tables from HTML Page}
\item{Metas}{Player Meta information (Name, Birth, Height, Weight, Shot, etc.)}
}
\description{
Get player info from a page
A function to manage scraping all of the info from Hockey-Reference.com player pages.
Also perform collection player meta info (name, birth, etc)
}
\keyword{internal}

