% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeRosters.R
\name{scrapeRosters}
\alias{scrapeRosters}
\title{Scrape all roster data}
\usage{
scrapeRosters(data_dir = "./data/rosters/", ...)
}
\arguments{
\item{data_dir}{Directory to store data in}

\item{...}{Additional parameters to pass}
}
\value{
roster information data.frame
}
\description{
This is a one-command function to scrape all roster data available from DailyFaceoff.com. 
Also saves results to a dated .RDS file.
}
\examples{
\dontrun{
scrapeRosters()
}
}

