% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{scrapeScores}
\alias{scrapeScores}
\title{Scrape and clean all score data}
\usage{
scrapeScores(data_dir = "./data/scores/", ...)
}
\arguments{
\item{data_dir}{Directory to store data in.}

\item{...}{Additional parameters to pass}
}
\value{
scores information data.frame
}
\description{
This is a one-command function to scrape and clean all score data available from Hockey-Reference.com.
Also saves results to a dated .RDS file.
}
\examples{
\dontrun{
scrapeScores()
}
}

