% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{tieSort}
\alias{tieSort}
\title{Sorting out ties}
\usage{
tieSort(game)
}
\arguments{
\item{game}{The game to sort ties in}
}
\value{
elo score value
}
\description{
This function sorts scores (including ties) into [0..1] win ratio for elo
}
\keyword{internal}
