% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayers.R
\name{updatePlayers}
\alias{updatePlayers}
\title{Update player information (don't rescrape old players)}
\usage{
updatePlayers(data_dir = "./data/players/", ...)
}
\arguments{
\item{data_dir}{The data dir of stored player information}

\item{...}{Additional parameters to pass}
}
\description{
Update player information (don't rescrape old players)
}

