% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{updateScores}
\alias{updateScores}
\title{Update score information (don't rescrape old scores)}
\usage{
updateScores(data_dir = "./data/players/", ...)
}
\arguments{
\item{data_dir}{The data dir of stored score information}

\item{...}{Additional parameters to pass}
}
\description{
Update score information (don't rescrape old scores)
}

