% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePlayersQH.R
\name{scrapePlayers.QH}
\alias{scrapePlayers.QH}
\title{Scrape and clean all player data}
\usage{
scrapePlayers.QH(player_list, directory = "./data/players/", ...)
}
\arguments{
\item{player_list}{A player list of the type type created by \code{\link{getPlayerList.QH}}}

\item{directory}{Directory to store data}

\item{...}{Additional parameters to pass}
}
\value{
a list of three cleaned data.frames, containing
\item{PlayerStats}{Combined player statistics}
\item{GoalieStats}{Combined goalie statistics}
\item{PlayerMeta}{Meta statistics for all (goalies and players)}
}
\description{
This is a one-command function to scrape and clean all player data available from Quanthockey.com. This takes many hours.
Also saves results to a dated .RDS file.
}
\examples{
\dontrun{scrapePlayers.QH()}
\dontrun{scrapePlayers.QH(sleep=15, long_sleep=180, group_by=500)}
}

