% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeScores.R
\name{updateScores}
\alias{updateScores}
\title{Update scores information (don't rescrape old scores).}
\usage{
updateScores(score_data, data_dir = "./data/scores/", ...)
}
\arguments{
\item{score_data}{The player_data data.frame to update}

\item{data_dir}{The data dir of stored player information}

\item{...}{Additional parameters to pass}
}
\value{
updated scores information data.frame
}
\description{
Update scores information (don't rescrape old scores).
}
\examples{
\dontrun{updateScores()}
\dontrun{updatePlayers(score_data, data_dir = './data/'}
}

